feat(progression): implement level-up event system and restore health mechanics
Rebalanced enemy stats (health, damage, attack speed, range) across all three tiers to improve gameplay difficulty curve. Updated weapon and passive item bonuses per level. Changed MoveSpeed from int to float for finer movement control. Added OnLevelUp event system to PlayerLevelManager and implemented health restoration to max when players level up. Removed unused imports and fixed code formatting.

+++ Assets/Scripts/Data/ScriptableObjects/EnemyTypeData_SO.cs
@@ -59,7 +59,7 @@ public struct EnemyTier
 public struct EnemyStats
 {
     public int MaxHealth;
-    public int MoveSpeed;
+    public float MoveSpeed;
     public int ATKDamage;
     public float ATKSpeed;
     public float ATKRange;
